BE-ApiGateway/src/main/java/com/wondoo/apigateway/utils/JwtUtils.java
package com.wondoo.apigateway.utils;

import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.SignatureAlgorithm;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

import javax.crypto.spec.SecretKeySpec;
import java.nio.charset.StandardCharsets;
import java.security.Key;

@Component
public class JwtUtils {
    private final Key key;

    public JwtUtils(@Value("${secret.key}") String secretKey) {
        this.key = new SecretKeySpec(secretKey.getBytes(StandardCharsets.UTF_8), SignatureAlgorithm.HS256.getJcaName());
    }

    public boolean validateJwt(String token) {
        String subject = getSubjectByJwt(token);
        return !(subject != null && !subject.isEmpty());
    }

    public String getSubjectByJwt(String token) {
        return Jwts.parserBuilder()
                .setSigningKey(key)
                .build()
                .parseClaimsJws(token)
                .getBody()
                .getSubject();
    }
}
